Added dynamic chapter num icon to both mobile and desktop
