+++ src/components/header/MobileHeader.jsx
@@ -1,14 +1,11 @@
 import logo from '../../assets/Website Assets - Phase 1/SVG/LOGO-header.svg';
 import { Link } from 'react-router-dom';
-import ComicPageAPI from '../../api/ComicPageAPI';
-import pageData from '../../api/data/pagesData.json';
-import MobileHeaderBackground from '../svgs/MobileHeaderBackground';
+import ChapterNumIcon from './ChapterNumIcon';
 import HamburgerMenu from './HamburgerMenu';
 
 import {COMIC_VIEWER_DEFAULT_PATH} from '../Main';
 
 export default function MobileHeader(props) {
-    let volNum = ComicPageAPI.getSeasonNum(pageData.maxDisplayPage);
 
     return (
         <div>
@@ -16,17 +13,9 @@ export default function MobileHeader(props) {
                 <div className="volDisplayMobile text-mocha-dark font-medium text-left ml-6 self-start">
                     <div className="volTextContainer flex flex-row w-full pt-3">
                         <div className="volText headerVolNum headerVolMobile">
-                            vol
+                            chap.
                         </div>
-                        <div className="volNumContainer ml-2">
-                            <svg width="32px" height="32px">
-                                <g>
-                                    <circle cx="14" cy="14"  r="13" strokeWidth="2" stroke="#998f7e" fillOpacity="0%"></circle>
-                                    <text x="14" y="14" fill="#998f7e" className="headerVolNum" textAnchor="middle" alignmentBaseline="central">{volNum}</text>
-                                </g>
-                            </svg>
-                        </div>
-                        
+                        <ChapterNumIcon />                        
                     </div>
                     
                 </div>
